Updated C# IDS schema automation
This commit contains the code that uses the c# code generation to convert the
legacy documentation testcases, so that they structurally comply with version
0.9.7 of the ids.xsd.
It relies on OldIds.cs class to read the legacy files, then converts them to
current ids.

the XSD tool is updated to add nullable attributes to C# classes.